+++ src/Event/Event.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/php-tools
  * @license     https://opensource.org/licenses/mit-license.php MIT License
- * @since       1.3.5
+ * @since       1.4.0
  */
 namespace Tools\Event;